Move userMode() && userModePrefix() to BufferState.js
